:sparkles: Made music transmission working

+++ src/main/kotlin/myra/bot/voice/voice/gateway/models/HelloPayload.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class HelloPayload(
-    @SerialName("heartbeat_interval") val heartbeatInterval: Long
+    @SerialName("heartbeat_interval") val heartbeatInterval: Double
 )
